backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
httpx
python-dotenv